+++ index.html
@@ -67,7 +67,6 @@
   <footer class="foota">
     <div>
       <span>Copyright ©️ 2022</span>
-      <span>&</span>
       <span>
         HARRY
       </span>
