+++ index.js
@@ -1,92 +1,89 @@
-// import allLibrary from './modules/switch.js';
 import './modules/date.js';
+import SwitchPage from './modules/switch.js';
+import allLibrary from './modules/class.js';
 
-const bookList = document.getElementById('book-list');
 const bookTitle = document.getElementById('title');
 const bookAuthor = document.getElementById('author');
 const bookAdd = document.getElementById('book-add');
 
-class Booklibrary {
-  constructor() {
-    this.books = [
-      {
-        title: 'Ghost',
-        author: 'Sidney Sheldon',
-      },
-      {
-        title: 'The Naked Face',
-        author: 'Sidney Sheldon',
-      },
-    ];
-  }
-  // add books
 
-  addBook(title, author) {
-    this.books.push({
-      title,
-      author,
-    });
+//   constructor() {
+//     this.books = [
+//       {
+//         title: 'Ghost',
+//         author: 'Sidney Sheldon',
+//       },
+//       {
+//         title: 'The Naked Face',
+//         author: 'Sidney Sheldon',
+//       },
+//     ];
+//   }
+//   // add books
 
-    bookAuthor.value = '';
-    bookTitle.value = '';
-    this.setLocalStorage();
-    this.displayBooks();
-  }
+//   addBook(title, author) {
+//     this.books.push({
+//       title,
+//       author,
+//     });
 
-  removeBook(index) {
-    this.books.splice(index, 1);
-    this.setLocalStorage();
-    this.displayBooks();
-  }
+//     bookAuthor.value = '';
+//     bookTitle.value = '';
+//     this.setLocalStorage();
+//     this.displayBooks();
+//   }
 
-  setLocalStorage() {
-    localStorage.setItem('localLibraries', JSON.stringify(this.books));
-  }
+//   removeBook(index) {
+//     this.books.splice(index, 1);
+//     this.setLocalStorage();
+//     this.displayBooks();
+//   }
 
-  getLocalStorage() {
-    if (localStorage.getItem('localLibraries')) {
-      this.books = JSON.parse(localStorage.getItem('localLibraries'));
-    }
-    this.displayBooks();
-  }
+//   setLocalStorage() {
+//     localStorage.setItem('localLibraries', JSON.stringify(this.books));
+//   }
 
-  displayBooks() {
-    bookList.innerHTML = '';
-    this.books.forEach((book, index) => {
-      const h3 = document.createElement('div');
-      h3.classList.add('book-div');
-      const bookNames = document.createElement('p');
-      const removeBtn = document.createElement('button');
-      bookNames.textContent = ` "${book.title}" by ${book.author}`;
+//   getLocalStorage() {
+//     if (localStorage.getItem('localLibraries')) {
+//       this.books = JSON.parse(localStorage.getItem('localLibraries'));
+//     }
+//     this.displayBooks();
+//   }
 
-      removeBtn.textContent = 'Remove';
-      removeBtn.classList.add('remove');
-      h3.append(bookNames, removeBtn);
-      removeBtn.addEventListener('click', () => {
-        this.removeBook(index);
-      });
-      bookList.appendChild(h3);
-    });
-  }
-}
+//   displayBooks() {
+//     bookList.innerHTML = '';
+//     this.books.forEach((book, index) => {
+//       const h3 = document.createElement('div');
+//       h3.classList.add('book-div');
+//       const bookNames = document.createElement('p');
+//       const removeBtn = document.createElement('button');
+//       bookNames.textContent = ` "${book.title}" by ${book.author}`;
 
-const allLibrary = new Booklibrary();
+//       removeBtn.textContent = 'Remove';
+//       removeBtn.classList.add('remove');
+//       h3.append(bookNames, removeBtn);
+//       removeBtn.addEventListener('click', () => {
+//         this.removeBook(index);
+//       });
+//       bookList.appendChild(h3);
+//     });
+//   }
+// }
+
+// const allLibrary = new Booklibrary();
 bookAdd.addEventListener('click', (e) => {
   e.preventDefault();
   allLibrary.addBook(bookTitle.value, bookAuthor.value);
+  allLibrary.displayBooks();
+  bookAuthor.value = '';
+  bookTitle.value = '';
 });
 
 document.addEventListener('DOMContentLoaded', () => {
   allLibrary.getLocalStorage();
+ 
 });
 
-const SwitchPage = (pageId) => {
-  const currentPage = document.querySelector('.pages .page.is-active');
-  currentPage.classList.remove('is-active');
-
-  const nextPage = document.querySelector(`.pages .page[data-page="${pageId}"]`);
-  nextPage.classList.add('is-active');
-};
 
 window.onload = () => {
   const tabSwitchers = document.querySelectorAll('[data-switcher]');
